# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AwaazPe API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface STTRequest {
  /** Base64 encoded audio data */
  audio: string;
  /** Language code (e.g. hi-IN, en-IN) */
  languageCode: string;
}

export interface STTResponse {
  transcript: string;
}

export type ParseRequestContactsItem = {
  name: string;
  upiId: string;
};

export interface ParseRequest {
  transcript: string;
  contacts?: ParseRequestContactsItem[];
}

export type ParseResponseAction =
  (typeof ParseResponseAction)[keyof typeof ParseResponseAction];

export const ParseResponseAction = {
  send: "send",
  schedule: "schedule",
  check_balance: "check_balance",
  history: "history",
  unknown: "unknown",
} as const;

export interface ParseResponse {
  action: ParseResponseAction;
  amount?: number | null;
  recipient?: string | null;
  recipientUpiId?: string | null;
  scheduledDate?: string | null;
  confidence: number;
  rawTranscript: string;
}

export interface TTSRequest {
  text: string;
  languageCode: string;
}

export interface TTSResponse {
  /** Base64 encoded audio */
  audio: string;
}

export interface TranslateRequest {
  text: string;
  sourceLanguage: string;
  targetLanguage: string;
}

export interface TranslateResponse {
  translatedText: string;
}