# lib/api-zod/src/generated/types/parseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AwaazPe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParseRequestContactsItem } from "./parseRequestContactsItem";

export interface ParseRequest {
  transcript: string;
  contacts?: ParseRequestContactsItem[];
}