# lib/api-zod/src/generated/types/parseResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AwaazPe API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParseResponseAction } from "./parseResponseAction";

export interface ParseResponse {
  action: ParseResponseAction;
  amount?: number | null;
  recipient?: string | null;
  recipientUpiId?: string | null;
  scheduledDate?: string | null;
  confidence: number;
  rawTranscript: string;
}