# lib/api-zod/src/generated/types/sTTRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AwaazPe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface STTRequest {
  /** Base64 encoded audio data */
  audio: string;
  /** Language code (e.g. hi-IN, en-IN) */
  languageCode: string;
}