# lib/api-zod/src/generated/types/tTSRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AwaazPe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TTSRequest {
  text: string;
  languageCode: string;
}