# lib/api-zod/src/generated/types/tTSResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AwaazPe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TTSResponse {
  /** Base64 encoded audio */
  audio: string;
}