# lib/api-zod/src/generated/types/translateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AwaazPe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranslateRequest {
  text: string;
  sourceLanguage: string;
  targetLanguage: string;
}