# lib/api-zod/src/generated/types/translateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AwaazPe API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranslateResponse {
  translatedText: string;
}